test: isolate migration error formatting
Move migration error formatting into a side-effect-free module so unit tests do not load dotenv or application configuration.


Copilot-Session: 25e0a7c2-d13b-4596-8100-df7bafb6e919
